+++ CHANGELOG.md
@@ -6,6 +6,12 @@ View all releases at: https://github.com/trimble-oss/modus-web-components/releas
 
 ## Unreleased
 
+### Added 
+
+- Added optional property `fullWidth` for modus-tabs. Now component can be resizable from UI if `fullWidth` flag is set on true.
+
+## 0.1.23 - 2022-10-26
+
 ### Added
 
 - Added a theme toggler to the toolbar in storybook pages to switch between `light` and `dark` themes